Add libfat
This is a library for direct access to a FAT filesystem.

I wrote it long ago for an university project and then rewrote it so
that I could put it on my resume (yes, back then I thought people
actually read the code). But looks like I forgot to create a repo when I
did that.

(Who cares about all that? I care. This is a personal archive for stuff
I don't need, but don't want to lose. It is here to reminisce about the
old days)